+++ test/style/blocks_test.exs
@@ -319,6 +319,59 @@ defmodule Styler.Style.BlocksTest do
       )
     end
 
+    test "transforms a `with` all the way to an `if` if necessary" do
+      # with a preroll
+      assert_style """
+      with foo <- bar,
+        true <- bop do
+        :ok
+      else
+        _ -> :error
+      end
+      """,
+      """
+      foo = bar
+
+      if bop do
+        :ok
+      else
+        :error
+      end
+      """
+      # no pre or postroll
+      assert_style """
+      with true <- bop do
+        :ok
+      else
+        _ -> :error
+      end
+      """,
+      """
+      if bop do
+        :ok
+      else
+        :error
+      end
+      """
+
+      # with postroll
+      assert_style """
+      with true <- bop, foo <- bar do
+        :ok
+      else
+        _ -> :error
+      end
+      """,
+      """
+      if bop do
+        foo = bar
+        :ok
+      else
+        :error
+      end
+      """
+    end
+
     test "moves non-arrow clauses from the beginning & end" do
       assert_style(
         """