+++ CHANGELOG.md
@@ -4,7 +4,8 @@
 
 ### Improvements
 
-* Rewrite `{Map|Keyword}.merge(single_key: value)` to use `put/3` instead
+* Rewrite `{Map|Keyword}.merge(single_key: value)` to use `put/3` instead (#96)
+* Attempt to keep comments in logical places when rewriting trivial `case` and `cond` statments (#97)
 
 ## v0.10.5