+++ CHANGELOG.md
@@ -4,11 +4,19 @@
 
 ### Improvements
 
-* `with`: rewrite trivial `lhs <- rhs` to `lhs = rhs` (#86)
-* `with`: rewrite with statements if statements when appropriate
-* `with`: switch keyword do to do block when adding clauses to the with body (`, do:` => `do end`)
+#### With Statements
+
+A slew of improvements for the `with` statement in this release:
+
+* rewrite trivial `lhs <- rhs` to `lhs = rhs` (#86)
+* rewrite with statements if statements when appropriate
+* switch keyword do to do block when adding clauses to the with body (`, do:` => `do end`)
+* put more effort into keeping comments in the right spot
+
+#### Other
+
 * Rewrite `{Map|Keyword}.merge(single_key: value)` to use `put/3` instead (#96)
-* Attempt to keep comments in logical places when rewriting trivial `case` and `cond` statments (#97)
+* Attempt to keep comments in logical places when rewriting trivial `case` and `cond` statements (#97)
 
 ## v0.10.5